Fixes #641 and the failing build.

+++ src/main/java/org/lisoft/lsml/model/item/BallisticWeapon.java
@@ -42,7 +42,7 @@ public class BallisticWeapon extends AmmoWeapon {
 
     public BallisticWeapon(
             // Item Arguments
-            String aName, String aDesc, String aMwoName, int aMwoId, int aSlots, double aTons, int aHP,
+            String aName, String aDesc, String aMwoName, int aMwoId, int aSlots, double aTons, double aHP,
             Faction aFaction,
             // HeatSource Arguments
             Attribute aHeat,
