+++ src/main/java/org/lisoft/lsml/model/item/EnergyWeapon.java
@@ -34,7 +34,7 @@ public class EnergyWeapon extends Weapon {
     protected final Attribute burnTime;
 
     public EnergyWeapon(// Item Arguments
-            String aName, String aDesc, String aMwoName, int aMwoId, int aSlots, double aTons, int aHP,
+            String aName, String aDesc, String aMwoName, int aMwoId, int aSlots, double aTons, double aHP,
             Faction aFaction,
             // HeatSource Arguments
             Attribute aHeat,
