+++ src/main/java/org/lisoft/lsml/model/item/TargetingComputer.java
@@ -38,7 +38,7 @@ public class TargetingComputer extends Module implements ModifierEquipment {
     private final List<Modifier> modifiers;
 
     public TargetingComputer(String aName, String aDesc, String aMwoName, int aMwoId, int aSlots, double aTons,
-            HardPointType aHardpointType, int aHP, Faction aFaction, List<Location> aAllowedLocations,
+            HardPointType aHardpointType, double aHP, Faction aFaction, List<Location> aAllowedLocations,
             List<ChassisClass> aAllowedChassisClasses, List<Modifier> aModifiers) {
         super(aName, aDesc, aMwoName, aMwoId, aSlots, aTons, aHardpointType, aHP, aFaction, aAllowedLocations,
                 aAllowedChassisClasses);
